+++ vcd/resource_vcd_vapp_vm.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/kr/pretty"
 	"github.com/vmware/go-vcloud-director/v2/govcd"
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
@@ -72,81 +73,527 @@ func resourceVcdVAppVm() *schema.Resource {
 	}
 }
 
-// getComputeValues
-// cpus, cores, memory, vmSizingPolicyId
-//
-// * CPU
-//   - vCpu Speed
-//   - vCpu Count
-//   - Cores per socket
-//   - CPU Reservation Guarantee
-//   - CPU Limit
-//   - CPU Shares
-//
-// * Memory
-//   - Memory MB
-//   - Memory Reservation Guarantee
-//   - Memory Limit
-//   - Memory Shares
-//
-// Priority comes from compute policy
-func getComputeValues(d *schema.ResourceData, vcdClient *VCDClient) (*int, *int, *int64, error) {
-	vdcComputePolicy, _, err := lookupComputePolicy(d, vcdClient)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("error finding sizing policy: %s", err)
+// VM Schema is defined as global so that it can be directly accessible in other places
+func vmSchemaFunc(vmType typeOfVm) map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"vapp_name": {
+			Type:        schema.TypeString,
+			Required:    vmType == vappVmType,
+			Optional:    vmType == standaloneVmType,
+			Computed:    vmType == standaloneVmType,
+			ForceNew:    vmType == vappVmType,
+			Description: "The vApp this VM belongs to - Required, unless it is a standalone VM",
+		},
+		"vm_type": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: fmt.Sprintf("Type of VM: either '%s' or '%s'", vappVmType, standaloneVmType),
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Required:    true,
+			ForceNew:    true,
+			Description: "A name for the VM, unique within the vApp",
+		},
+		"computer_name": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "Computer name to assign to this virtual machine",
+		},
+		"org": {
+			Type:     schema.TypeString,
+			Optional: true,
+			ForceNew: true,
+			Description: "The name of organization to use, optional if defined at provider " +
+				"level. Useful when connected as sysadmin working across different organizations",
+		},
+		"vdc": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			ForceNew:    true,
+			Description: "The name of VDC to use, optional if defined at provider level",
+		},
+		"template_name": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			ForceNew:    true,
+			Description: "The name of the vApp Template to use",
+			//RequiredWith: []string{"template_name", "catalog_name"},
+		},
+		"vm_name_in_template": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			ForceNew:    true,
+			Description: "The name of the VM in vApp Template to use. In cases when vApp template has more than one VM",
+			//RequiredWith: []string{"vm_name_in_template", "catalog_name", "template_name"},
+		},
+		"catalog_name": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "The catalog name in which to find the given vApp Template or media for boot_image",
+			//RequiredWith: []string{"catalog_name", "template_name", "boot_image"},
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "The VM description",
+		},
+		"memory": {
+			Type:         schema.TypeInt,
+			Optional:     true,
+			Computed:     true,
+			Description:  "The amount of RAM (in MB) to allocate to the VM",
+			ValidateFunc: validateMultipleOf4(),
+		},
+		"memory_reservation": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The amount of RAM (in MB) reservation on the underlying virtualization infrastructure",
+		},
+		"memory_priority": {
+			Type:         schema.TypeString,
+			Optional:     true,
+			Computed:     true,
+			Description:  "Pre-determined relative priorities according to which the non-reserved portion of this resource is made available to the virtualized workload",
+			ValidateFunc: validation.StringInSlice([]string{"LOW", "NORMAL", "HIGH", "CUSTOM"}, false),
+		},
+		"memory_shares": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "Custom priority for the resource. This is a read-only, unless the `memory_priority` is CUSTOM",
+		},
+		"memory_limit": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The limit for how much of memory can be consumed on the underlying virtualization infrastructure. This is only valid when the resource allocation is not unlimited.",
+		},
+		"cpus": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The number of virtual CPUs to allocate to the VM",
+		},
+		"cpu_cores": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The number of cores per socket",
+		},
+		"cpu_reservation": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The amount of MHz reservation on the underlying virtualization infrastructure",
+		},
+		"cpu_priority": {
+			Type:         schema.TypeString,
+			Optional:     true,
+			Computed:     true,
+			Description:  "Pre-determined relative priorities according to which the non-reserved portion of this resource is made available to the virtualized workload",
+			ValidateFunc: validation.StringInSlice([]string{"LOW", "NORMAL", "HIGH", "CUSTOM"}, false),
+		},
+		"cpu_shares": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "Custom priority for the resource. This is a read-only, unless the `cpu_priority` is CUSTOM",
+		},
+		"cpu_limit": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The limit for how much of CPU can be consumed on the underlying virtualization infrastructure. This is only valid when the resource allocation is not unlimited.",
+		},
+		"metadata": {
+			Type:     schema.TypeMap,
+			Optional: true,
+			// For now underlying go-vcloud-director repo only supports
+			// a value of type String in this map.
+			Description: "Key value map of metadata to assign to this VM",
+		},
+		"href": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "VM Hyper Reference",
+		},
+		"accept_all_eulas": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     true,
+			Description: "Automatically accept EULA if OVA has it",
+		},
+		"power_on": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     true,
+			Description: "A boolean value stating if this VM should be powered on",
+		},
+		"storage_profile": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "Storage profile to override the default one",
+		},
+		"os_type": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "Operating System type. Possible values can be found in documentation.",
+		},
+		"hardware_version": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "Virtual Hardware Version (e.g.`vmx-14`, `vmx-13`, `vmx-12`, etc.)",
+		},
+		"boot_image": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Media name to add as boot image.",
+			//RequiredWith: []string{"boot_image", "catalog_name"},
+		},
+		"network_dhcp_wait_seconds": {
+			Optional:     true,
+			Type:         schema.TypeInt,
+			ValidateFunc: validation.IntAtLeast(0),
+			Description: "Optional number of seconds to try and wait for DHCP IP (valid for " +
+				"'network' block only)",
+		},
+		"network": {
+			Optional:    true,
+			Type:        schema.TypeList,
+			Description: " A block to define network interface. Multiple can be used.",
+			Elem: &schema.Resource{
+				Schema: map[string]*schema.Schema{
+					"type": {
+						Required:     true,
+						Type:         schema.TypeString,
+						ValidateFunc: vmNetworkTypeValidator(vmType),
+						Description:  "Network type to use: 'vapp', 'org' or 'none'. Use 'vapp' for vApp network, 'org' to attach Org VDC network. 'none' for empty NIC.",
+					},
+					"ip_allocation_mode": {
+						Optional:     true,
+						Type:         schema.TypeString,
+						ValidateFunc: validation.StringInSlice([]string{"POOL", "DHCP", "MANUAL", "NONE"}, false),
+						Description:  "IP address allocation mode. One of POOL, DHCP, MANUAL, NONE",
+					},
+					"name": {
+						ForceNew:    false,
+						Optional:    true, // In case of type = none it is not required
+						Type:        schema.TypeString,
+						Description: "Name of the network this VM should connect to. Always required except for `type` `NONE`",
+					},
+					"ip": {
+						Computed:     true,
+						Optional:     true,
+						Type:         schema.TypeString,
+						ValidateFunc: checkEmptyOrSingleIP(), // Must accept empty string to ease using HCL interpolation
+						Description:  "IP of the VM. Settings depend on `ip_allocation_mode`. Omitted or empty for DHCP, POOL, NONE. Required for MANUAL",
+					},
+					"is_primary": {
+						Optional: true,
+						Computed: true,
+						// By default, if the value is omitted it will report schema change
+						// on every terraform operation. The below function
+						// suppresses such cases "" => "false" when applying.
+						DiffSuppressFunc: falseBoolSuppress(),
+						Type:             schema.TypeBool,
+						Description:      "Set to true if network interface should be primary. First network card in the list will be primary by default",
+					},
+					"mac": {
+						Computed:    true,
+						Optional:    true,
+						Type:        schema.TypeString,
+						Description: "Mac address of network interface",
+					},
+					"adapter_type": {
+						Type:             schema.TypeString,
+						Computed:         true,
+						Optional:         true,
+						DiffSuppressFunc: suppressCase,
+						Description:      "Network card adapter type. (e.g. 'E1000', 'E1000E', 'SRIOVETHERNETCARD', 'VMXNET3', 'PCNet32')",
+					},
+					"connected": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Default:     true,
+						Description: "It defines if NIC is connected or not.",
+					},
+				},
+			},
+		},
+		"disk": {
+			Type: schema.TypeSet,
+			Elem: &schema.Resource{Schema: map[string]*schema.Schema{
+				"name": {
+					Type:        schema.TypeString,
+					Required:    true,
+					Description: "Independent disk name",
+				},
+				"bus_number": {
+					Type:        schema.TypeString,
+					Required:    true,
+					Description: "Bus number on which to place the disk controller",
+				},
+				"unit_number": {
+					Type:        schema.TypeString,
+					Required:    true,
+					Description: "Unit number (slot) on the bus specified by BusNumber",
+				},
+				"size_in_mb": {
+					Type:        schema.TypeInt,
+					Computed:    true,
+					Description: "The size of the disk in MB.",
+				},
+			}},
+			Optional: true,
+			Set:      resourceVcdVmIndependentDiskHash,
+		},
+		"override_template_disk": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			ForceNew:    true,
+			Description: "A block to match internal_disk interface in template. Multiple can be used. Disk will be matched by bus_type, bus_number and unit_number.",
+			Elem: &schema.Resource{Schema: map[string]*schema.Schema{
+				"bus_type": {
+					Type:         schema.TypeString,
+					Required:     true,
+					ForceNew:     true,
+					ValidateFunc: validation.StringInSlice([]string{"ide", "parallel", "sas", "paravirtual", "sata", "nvme"}, false),
+					Description:  "The type of disk controller. Possible values: ide, parallel( LSI Logic Parallel SCSI), sas(LSI Logic SAS (SCSI)), paravirtual(Paravirtual (SCSI)), sata, nvme",
+				},
+				"size_in_mb": {
+					Type:        schema.TypeInt,
+					ForceNew:    true,
+					Required:    true,
+					Description: "The size of the disk in MB.",
+				},
+				"bus_number": {
+					Type:        schema.TypeInt,
+					ForceNew:    true,
+					Required:    true,
+					Description: "The number of the SCSI or IDE controller itself.",
+				},
+				"unit_number": {
+					Type:        schema.TypeInt,
+					ForceNew:    true,
+					Required:    true,
+					Description: "The device number on the SCSI or IDE controller of the disk.",
+				},
+				"iops": {
+					Type:        schema.TypeInt,
+					ForceNew:    true,
+					Optional:    true,
+					Description: "Specifies the IOPS for the disk. Default is 0.",
+				},
+				"storage_profile": {
+					Type:        schema.TypeString,
+					ForceNew:    true,
+					Optional:    true,
+					Description: "Storage profile to override the VM default one",
+				},
+			}},
+		},
+		"internal_disk": {
+			Type:        schema.TypeList,
+			Computed:    true,
+			Description: "A block will show internal disk details",
+			Elem: &schema.Resource{Schema: map[string]*schema.Schema{
+				"disk_id": {
+					Type:        schema.TypeString,
+					Computed:    true,
+					Description: "The disk ID.",
+				},
+				"bus_type": {
+					Type:        schema.TypeString,
+					Computed:    true,
+					Description: "The type of disk controller. Possible values: ide, parallel( LSI Logic Parallel SCSI), sas(LSI Logic SAS (SCSI)), paravirtual(Paravirtual (SCSI)), sata, nvme",
+				},
+				"size_in_mb": {
+					Type:        schema.TypeInt,
+					Computed:    true,
+					Description: "The size of the disk in MB.",
+				},
+				"bus_number": {
+					Type:        schema.TypeInt,
+					Computed:    true,
+					Description: "The number of the SCSI or IDE controller itself.",
+				},
+				"unit_number": {
+					Type:        schema.TypeInt,
+					Computed:    true,
+					Description: "The device number on the SCSI or IDE controller of the disk.",
+				},
+				"thin_provisioned": {
+					Type:        schema.TypeBool,
+					Computed:    true,
+					Description: "Specifies whether the disk storage is pre-allocated or allocated on demand.",
+				},
+				"iops": {
+					Type:        schema.TypeInt,
+					Computed:    true,
+					Description: "Specifies the IOPS for the disk. Default is 0.",
+				},
+				"storage_profile": {
+					Type:        schema.TypeString,
+					Computed:    true,
+					Description: "Storage profile to override the VM default one",
+				},
+			}},
+		},
+		"expose_hardware_virtualization": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     false,
+			Description: "Expose hardware-assisted CPU virtualization to guest OS.",
+		},
+		"guest_properties": {
+			Type:        schema.TypeMap,
+			Optional:    true,
+			Description: "Key/value settings for guest properties",
+		},
+		"customization": {
+			Optional:    true,
+			Computed:    true,
+			MinItems:    1,
+			MaxItems:    1,
+			Type:        schema.TypeList,
+			Description: "Guest customization block",
+			Elem: &schema.Resource{
+				Schema: map[string]*schema.Schema{
+					"force": {
+						ValidateFunc: noopValueWarningValidator(true,
+							"Using 'true' value for field 'vcd_vapp_vm.customization.force' will reboot VM on every 'terraform apply' operation"),
+						Type:     schema.TypeBool,
+						Optional: true,
+						Default:  false,
+						// This settings is used as a 'flag' and it does not matter what is set in the
+						// state. If it is 'true' - then it means that 'update' procedure must set the
+						// VM for customization at next boot and reboot it.
+						DiffSuppressFunc: suppressFalse(),
+						Description:      "'true' value will cause the VM to reboot on every 'apply' operation",
+					},
+					"enabled": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "'true' value will enable guest customization. It may occur on first boot or when 'force' is used",
+					},
+					"change_sid": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "'true' value will change SID. Applicable only for Windows VMs",
+					},
+					"allow_local_admin_password": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "Allow local administrator password",
+					},
+					"must_change_password_on_first_login": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "Require Administrator to change password on first login",
+					},
+					"auto_generate_password": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "Auto generate password",
+					},
+					"admin_password": {
+						Type:        schema.TypeString,
+						Optional:    true,
+						Computed:    true,
+						Sensitive:   true,
+						Description: "Manually specify admin password",
+					},
+					"number_of_auto_logons": {
+						Type:         schema.TypeInt,
+						Optional:     true,
+						Computed:     true,
+						Description:  "Number of times to log on automatically. '0' - disabled.",
+						ValidateFunc: validation.IntAtLeast(0),
+					},
+					"join_domain": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "Enable this VM to join a domain",
+					},
+					"join_org_domain": {
+						Type:        schema.TypeBool,
+						Optional:    true,
+						Computed:    true,
+						Description: "Use organization's domain for joining",
+					},
+					"join_domain_name": {
+						Type:        schema.TypeString,
+						Optional:    true,
+						Computed:    true,
+						Description: "Custom domain name for join",
+					},
+					"join_domain_user": {
+						Type:        schema.TypeString,
+						Optional:    true,
+						Computed:    true,
+						Description: "Username for custom domain name join",
+					},
+					"join_domain_password": {
+						Type:        schema.TypeString,
+						Optional:    true,
+						Computed:    true,
+						Sensitive:   true,
+						Description: "Password for custom domain name join",
+					},
+					"join_domain_account_ou": {
+						Type:        schema.TypeString,
+						Optional:    true,
+						Computed:    true,
+						Description: "Account organizational unit for domain name join",
+					},
+					"initscript": {
+						Type:        schema.TypeString,
+						Optional:    true,
+						Computed:    true,
+						Description: "Script to run on initial boot or with customization.force=true set",
+					},
+				},
+			},
+		},
+		"cpu_hot_add_enabled": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     false,
+			Description: "True if the virtual machine supports addition of virtual CPUs while powered on.",
+		},
+		"memory_hot_add_enabled": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     false,
+			Description: "True if the virtual machine supports addition of memory while powered on.",
+		},
+		"prevent_update_power_off": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Default:     false,
+			Description: "True if the update of resource should fail when virtual machine power off needed.",
+		},
+		"sizing_policy_id": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "VM sizing policy ID. Has to be assigned to Org VDC.",
+		},
 	}
-
-	var (
-		setCpu    *int
-		setCores  *int
-		setMemory *int64
-	)
-
-	// If VDC Compute policy is not set - we're specifying CPU and Memory directly
-	// if vdcComputePolicy == nil {
-	if memory, isMemorySet := d.GetOk("memory"); isMemorySet {
-		memInt := int64(memory.(int))
-		setMemory = &memInt
-	}
-
-	if cpus, isCpusSet := d.GetOk("cpus"); isCpusSet {
-		cpuInt := cpus.(int)
-		setCpu = &cpuInt
-	}
-
-	if cpuCores, isCpuCoresSet := d.GetOk("cpu_cores"); isCpuCoresSet {
-		cpuCoresInt := cpuCores.(int)
-		setCores = &cpuCoresInt
-	}
-
-	// return setCpu, setCores, setMemory, nil
-
-	// }
-
-	// Check if sizing policy has any settings settings and override VM configuration with it
-
-	if vdcComputePolicy != nil {
-		if vdcComputePolicy.Memory != nil {
-			mem := int64(*vdcComputePolicy.Memory)
-			setMemory = &mem
-		}
-
-		if vdcComputePolicy.CPUCount != nil {
-			setCpu = vdcComputePolicy.CPUCount
-		}
-
-		if vdcComputePolicy.CoresPerSocket != nil {
-			setCores = vdcComputePolicy.CoresPerSocket
-		}
-
-		// memoryInComputePolicy := vdcComputePolicy.Memory
-		// cpuInComputePolicy := vdcComputePolicy.CPUCount
-		// cpuCores := vdcComputePolicy.CoresPerSocket
-	}
-
-	// Lookup parameters from VM resource schema
-
-	return setCpu, setCores, setMemory, nil
 }
 
 // resourceVcdVAppVmCreate is an entry function for VM within vApp creation. It locks parent vApp and cascades down the
@@ -295,46 +742,6 @@ func genericResourceVmCreate(d *schema.ResourceData, meta interface{}, vmType ty
 		//}
 		return diag.Errorf("error attaching-detaching  disks when updating resource : %s", err)
 	}
-
-	// Still to fix
-	///////////////////////
-	sizingPolicy, _, err := lookupComputePolicy(d, vcdClient)
-	if err != nil {
-		return diag.Errorf("error finding sizing policy: %s", err)
-	}
-	computePolicy := sizingPolicy
-
-	memory, isMemorySet := d.GetOk("memory")
-	isMemoryComingFromSizingPolicy := computePolicy != nil && (computePolicy.Memory != nil && !isMemorySet)
-	if isMemoryComingFromSizingPolicy && isMemorySet {
-		logForScreen("vcd_vapp_vm", fmt.Sprintf("WARNING: sizing policy is specifying a memory of %d that won't be overriden by `memory` attribute", *computePolicy.Memory))
-	}
-
-	if !isMemoryComingFromSizingPolicy {
-		err = vm.ChangeMemory(int64(memory.(int)))
-		if err != nil {
-			return diag.FromErr(err)
-		}
-	}
-
-	cpus, isCpusSet := d.GetOk("cpus")
-	cpuCores, isCpuCoresSet := d.GetOk("cpu_cores")
-	isCpuComingFromSizingPolicy := computePolicy != nil && ((computePolicy.CPUCount != nil && !isCpusSet) || (computePolicy.CoresPerSocket != nil && !isCpuCoresSet))
-	if isCpuComingFromSizingPolicy && isCpusSet {
-		logForScreen("vcd_vapp_vm", fmt.Sprintf("WARNING: sizing policy is specifying CPU count of %d that won't be overriden by `cpus` attribute", *computePolicy.CPUCount))
-	}
-	if isCpuComingFromSizingPolicy && isCpuCoresSet {
-		logForScreen("vcd_vapp_vm", fmt.Sprintf("WARNING: sizing policy is specifying %d CPU cores that won't be overriden by `cpu_cores` attribute", *computePolicy.CoresPerSocket))
-	}
-
-	if !isCpuComingFromSizingPolicy {
-		err = vm.ChangeCPU(cpus.(int), cpuCores.(int))
-		if err != nil {
-			return diag.FromErr(err)
-		}
-	}
-
-	///////////////////////////////
 
 	// Handle Advanced compute settings CPU and Memory shares, limits and reservation
 	// Such schema fields are processed:
@@ -554,7 +961,7 @@ func createVmFromTemplate(d *schema.ResourceData, meta interface{}, vmType typeO
 			Deploy:           false,
 			Name:             vapp.VApp.Name,
 			PowerOn:          false, // Power on is set to false as there will be additional operations before final VM creation
-			Description:      description,
+			Description:      vapp.VApp.Description,
 			SourcedItem: &types.SourcedCompositionItemParam{
 				Source: &types.Reference{
 					HREF: templateHref,
@@ -632,6 +1039,28 @@ func createVmFromTemplate(d *schema.ResourceData, meta interface{}, vmType typeO
 	err = updateHardwareVersionAndOsType(d, vm)
 	if err != nil {
 		return nil, fmt.Errorf("error updating hardware version and OS type : %s", err)
+	}
+
+	// vApp template VMs additionaly require CPU/Memory sorting if
+	// Lookup CPU values either from schema or from sizing policy. If nothing is set - it will be inherited from
+	cpuCores, cpuCoresPerSocket, memory, err := getComputeValues(d, vcdClient)
+	if err != nil {
+		return nil, fmt.Errorf("error getting CPU/Memory compute values: %s", err)
+	}
+
+	if cpuCores != nil {
+		err = vm.ChangeCPU(*cpuCores, *cpuCoresPerSocket)
+		if err != nil {
+			return nil, fmt.Errorf("error changing CPU settings: %s", err)
+		}
+
+	}
+
+	if memory != nil {
+		err = vm.ChangeMemory(*memory)
+		if err != nil {
+			return nil, fmt.Errorf("error setting memory size from schema for VM from template: %s", err)
+		}
 	}
 
 	return vm, nil
@@ -814,13 +1243,6 @@ func createVmEmpty(d *schema.ResourceData, meta interface{}, vmType typeOfVm) (*
 			params.CreateVm.VmSpecSection.MemoryResourceMb = &types.MemoryResourceMb{Configured: *memory}
 		}
 
-		// If compute policy is nil, CPU and memory settings must be applied on create to avoid API error
-		// if vmComputePolicy == nil {
-		// 	params.CreateVm.VmSpecSection.NumCpus = takeIntPointer(d.Get("cpus").(int))
-		// 	params.CreateVm.VmSpecSection.NumCoresPerSocket = takeIntPointer(d.Get("cpu_cores").(int))
-		// 	params.CreateVm.VmSpecSection.MemoryResourceMb = &types.MemoryResourceMb{Configured: int64(d.Get("sizing_policy_id").(int))}
-		// }
-
 		newVm, err = vdc.CreateStandaloneVm(&params)
 		if err != nil {
 			return nil, err
@@ -879,10 +1301,6 @@ func createVmEmpty(d *schema.ResourceData, meta interface{}, vmType typeOfVm) (*
 			recomposeVAppParamsForEmptyVm.CreateItem.VmSpecSection.MemoryResourceMb = &types.MemoryResourceMb{Configured: *memory}
 		}
 
-		// err = addSizingPolicy(d, vcdClient, recomposeVAppParamsForEmptyVm)
-		// if err != nil {
-		// 	return nil, err
-		// }
 		util.Logger.Printf("[VM create - add empty VM] recomposeVAppParamsForEmptyVm %# v", pretty.Formatter(recomposeVAppParamsForEmptyVm))
 		newVm, err = vapp.AddEmptyVm(recomposeVAppParamsForEmptyVm)
 		if err != nil {
@@ -893,90 +1311,6 @@ func createVmEmpty(d *schema.ResourceData, meta interface{}, vmType typeOfVm) (*
 	default:
 		return nil, fmt.Errorf("unknown VM type %s", vmType)
 	}
-
-	//recomposeVAppParamsForEmptyVm := &types.RecomposeVAppParamsForEmptyVm{
-	//	XmlnsVcloud: types.XMLNamespaceVCloud,
-	//	XmlnsOvf:    types.XMLNamespaceOVF,
-	//	PowerOn:     false, // Power on is handled at the end of VM creation process
-	//	CreateItem: &types.CreateItem{
-	//		Name: vmName,
-	//		// Bug in vCD - accepts only org VDC networks and automatically add them. We add them with update
-	//		// BUG in vCD 9.5 version, do not allow empty NetworkConnectionSection, so we pass simplest network configuration
-	//		// and after VM created update with real config
-	//		NetworkConnectionSection: &types.NetworkConnectionSection{
-	//			PrimaryNetworkConnectionIndex: 0,
-	//			NetworkConnection: []*types.NetworkConnection{
-	//				{Network: "none", NetworkConnectionIndex: 0, IPAddress: "any", IsConnected: false, IPAddressAllocationMode: "NONE"}},
-	//		},
-	//		StorageProfile:            storageProfilePtr,
-	//		Description:               d.Get("description").(string),
-	//		GuestCustomizationSection: customizationSection,
-	//		VmSpecSection: &types.VmSpecSection{
-	//			Modified:          takeBoolPointer(true),
-	//			Info:              "Virtual Machine specification",
-	//			OsType:            osType.(string),
-	//			NumCpus:           takeIntPointer(d.Get("cpus").(int)),
-	//			NumCoresPerSocket: takeIntPointer(d.Get("cpu_cores").(int)),
-	//			CpuResourceMhz:    &types.CpuResourceMhz{Configured: 0},
-	//			MemoryResourceMb:  &types.MemoryResourceMb{Configured: int64(memory.(int))},
-	//			MediaSection:      nil,
-	//			// can be created with resource internal_disk
-	//			DiskSection:      &types.DiskSection{DiskSettings: []*types.DiskSettings{}},
-	//			HardwareVersion:  &types.HardwareVersion{Value: hardWareVersion.(string)}, // need support older version vCD
-	//			VmToolsVersion:   "",
-	//			VirtualCpuType:   virtualCpuType,
-	//			TimeSyncWithHost: nil,
-	//		},
-	//		BootImage: bootImage,
-	//	},
-	//	AllEULAsAccepted: true,
-	//}
-	//
-	//err = addSizingPolicy(d, vcdClient, recomposeVAppParamsForEmptyVm)
-	//if err != nil {
-	//	return nil, err
-	//}
-	//
-	//var mediaReference *types.Reference
-	//if bootImage != nil {
-	//	mediaReference = &types.Reference{
-	//		HREF: bootImage.HREF,
-	//		ID:   bootImage.ID,
-	//		Type: bootImage.Type,
-	//		Name: bootImage.Name,
-	//	}
-	//}
-	//// Standalone VM
-	//if vappName == "" {
-	//	params := types.CreateVmParams{
-	//		Xmlns:       types.XMLNamespaceVCloud,
-	//		Name:        vmName,
-	//		PowerOn:     false,
-	//		Description: recomposeVAppParamsForEmptyVm.CreateItem.Description,
-	//		CreateVm: &types.Vm{
-	//			Name:                      vmName,
-	//			ComputePolicy:             recomposeVAppParamsForEmptyVm.CreateItem.ComputePolicy,
-	//			NetworkConnectionSection:  recomposeVAppParamsForEmptyVm.CreateItem.NetworkConnectionSection,
-	//			VmSpecSection:             recomposeVAppParamsForEmptyVm.CreateItem.VmSpecSection,
-	//			GuestCustomizationSection: recomposeVAppParamsForEmptyVm.CreateItem.GuestCustomizationSection,
-	//			StorageProfile:            recomposeVAppParamsForEmptyVm.CreateItem.StorageProfile,
-	//		},
-	//		Media: mediaReference,
-	//	}
-	//	newVm, err = vdc.CreateStandaloneVm(&params)
-	//	if err != nil {
-	//		return nil, err
-	//	}
-	//} else {
-	//	util.Logger.Printf("[VM create - add empty VM] recomposeVAppParamsForEmptyVm %# v", pretty.Formatter(recomposeVAppParamsForEmptyVm))
-	//	newVm, err = vapp.AddEmptyVm(recomposeVAppParamsForEmptyVm)
-	//	if err != nil {
-	//		d.SetId("")
-	//		return nil, fmt.Errorf("[VM creation] error creating VM %s : %s", vmName, err)
-	//	}
-	//}
-	//
-	//d.SetId(newVm.VM.ID)
 
 	util.Logger.Printf("[VM create] VM after creation %# v", pretty.Formatter(newVm.VM))
 	vapp, err = newVm.GetParentVApp()
